+++ npm-publish-gradle-plugin/src/main/kotlin/util/NpmPublishDsl.kt
@@ -1,0 +1,5 @@
+package dev.petuska.npm.publish.util
+
+@Target(AnnotationTarget.CLASS, AnnotationTarget.TYPE)
+@DslMarker
+internal annotation class NpmPublishDsl