+++ npm-publish-gradle-plugin/src/main/kotlin/util/overrideMap.kt
@@ -8,10 +8,12 @@ internal fun MutableMap<String, Any>.overrideFrom(other: Map<String, Any>): Muta
       old == null -> new
       old is Map<*, *> && new is Map<*, *> -> (old as Map<String, Any>).toMutableMap()
         .overrideFrom(new as Map<String, Any>)
+
       old is Array<*> && new is Array<*> -> old.toList() + new.toList()
       old is Array<*> && new is Collection<*> -> old.toList() + new
       (old is Collection<*> && new is Array<*>) ||
         (old is Collection<*> && new is Collection<*>) -> old + new
+
       else -> new
     }
   }